+++ packages/cli/src/logger/__mocks__/logger.mock.ts
@@ -1,0 +1,9 @@
+import type { Writer } from "../types";
+
+import * as sinon from "sinon";
+
+export default class LoggerMock {
+  readonly debug: Writer = sinon.fake();
+  readonly info: Writer = sinon.fake();
+  readonly warn: Writer = sinon.fake();
+}