+++ packages/cli/src/logger/__mocks__/writer.mock.ts
@@ -1,0 +1,5 @@
+import * as sinon from "sinon";
+
+const WriterMock = sinon.fake();
+
+export default WriterMock;